+++ Demo/src/main/java/com/paypal/android/ui/paypalwebvault/PayPalWebVaultViewModel.kt
@@ -86,15 +86,13 @@ class PayPalWebVaultViewModel @Inject constructor(
     ) {
         vaultPayPalState = ActionState.Loading
 
-        paypalClient.vault(activity, request) { result ->
-            when (result) {
-                is PayPalPresentAuthChallengeResult.Success -> {
-                    // do nothing; wait for user to authenticate PayPal vault in Chrome Custom Tab
-                }
-
-                is PayPalPresentAuthChallengeResult.Failure ->
-                    vaultPayPalState = ActionState.Failure(result.error)
+        when (val result = paypalClient.vault(activity, request)) {
+            is PayPalPresentAuthChallengeResult.Success -> {
+                // do nothing; wait for user to authenticate PayPal vault in Chrome Custom Tab
             }
+
+            is PayPalPresentAuthChallengeResult.Failure ->
+                vaultPayPalState = ActionState.Failure(result.error)
         }
     }