+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## Unreleased
 
 * PayPalWebPayments
-  * updates client config in web checkout
-* Migrates to kotlin serialization for all data serialization
+  * Add `PayPalWebCheckoutClient.start(activity, request, callback)` method with asynchronous
+    callback support
+  * Deprecates `PayPalWebCheckoutClient.start(activity, request)`
 
 ## 2.2.0 (2025-10-01)
 * CardPayments